+++ CHANGELOG.md
@@ -10,6 +10,7 @@ v0.9.8-dev (August xx, 2021)
 * Add questions actions to redirect to project home page or project components.
 * Support auto start of project via global System Setting.
 * Create new route for displaying a single system component control.
+* New controls.models.System property producer_elements_control_impl_smts_dict to get dictionary of control implementation statements associated with a system element.
 
 **Data changes**